explicit c++-style casting from double to integral types (for older gccs)
git-svn-id: svn://svn.code.sf.net/p/stxxl/code/trunk@2074 4e380d45-d1fd-0310-85a7-d18ec86df0ad

+++ include/stxxl/bits/common/utils.h
@@ -112,13 +112,13 @@ STXXL_MAX(const Tp & a, const Tp & b)
 template <typename Integer>
 inline Integer log2_ceil(Integer i)
 {
-    return ceil(log2(i));
+    return Integer(ceil(log2(i)));
 }
 
 template <typename Integer>
 inline Integer log2_floor(Integer i)
 {
-    return log2(i);
+    return Integer(log2(i));
 }
 
 ////////////////////////////////////////////////////////////////////////////
